MAINT: Updated author headers. Renamed FiffRawEvent to AnnotationView
