+++ applications/mne_analyze/libs/anShared/Model/fiffrawviewmodel.h
@@ -4,6 +4,7 @@
  * @author   Lorenz Esch <lesch@mgh.harvard.edu>;
  *           Lars Debor <Lars.Debor@tu-ilmenau.de>;
  *           Simon Heinke <Simon.Heinke@tu-ilmenau.de>
+ *           Gabriel Motta <gbmotta@mgh.harvard.edu>
  * @version  dev
  * @date     October, 2018
  *
